requirements_moviepy.txt
# Requirements para versão MoviePy
Pillow>=10.0.0
pyinstaller>=6.0.0
moviepy>=1.0.3
numpy>=1.24.0
